1-20 【大更新】修复了登录和注册的不安全设计、调整表结构以及更多（主要集中于上传视频的更新） 1. 原来的登录/注册参数依靠param传递，这是不安全的，现在我们设计了DTO来封装这些信息，并将其保存在Body中传递。并且我们将登录接口从GET改回POST。现在admin和user登录均已改为更安全的POST+body传递DTO方式。 2. 调整了表结构。现在video_info_file(以及upload)的fileId都是Long，使用雪花算法生成。 3. 增加了删除视频文件时将删除路径传递给MQ，异步删除功能。在生产者端（niloweb/service/CreativeCenterVideoUploadService.java）和消费者端（nilomqconsumer/config/RabbitMqConfig.java）配置了RabbitMQ的序列化，强制使用JSON序列化。目前异步删除的操作仅在windows端测试通过。 4. 更新了StringUtil的isValidPath方法，现在这个方法修复了两个安全漏洞：     1. 修复了简单的前缀判断导致的漏洞问题（原来判断/opt/file前缀时，/opt/file_1也视为有效，现在利用NIO的API将其转化为真实目录来判断）     2. 修复了当程序运行在windows下由于windows将“\”也视作目录分割线导致的分割错误的问题。现在通过将"\"转化为"/"来统一处理 5. 调整common包的自动配置，现在部分自动配置会在没有导入redis以来的情况下避免自动配置引用redis的自动配置类。 6. 现在配置了rabbitmq和consumer的重试机制，consumer在spring框架端配置了消费者程序抛出异常本地重试3次机制，如果还是出异常自动向MQ提交nack，保证最后都会提交nack。 7. 现在确定下来VideoInfoFileUpload这张表保存文件路径的策略了，现在只保存相对路径，原因有二：一是防止前端泄露服务器文件结构，二是防止数据库内容迁移因为OS不同导致文件路径全部失效。 8. 创建了一个消费者类，用于消费MQ的消息 9. 视频转码部分完成，发送到MQ之前的部分已经完成，但是消费者端还差点

+++ nilo-common/src/main/java/com/neon/nilocommon/entity/po/VideoInfoUpload.java
@@ -58,12 +58,12 @@ public class VideoInfoUpload
     /**
      * 0:转码中 1:转码失败 2:待审核 3:审核成功 4:审核失败
      */
-    private Integer status;
+    private Short status;
 
     /**
      * 0:自制作 1:转载
      */
-    private Integer postType;
+    private Short postType;
 
     /**
      * 原资源说明
